Move getArgs out of parse
